Address feedback: Move util functions to `util/alias.js`
