+++ src/util.js
@@ -1,2 +1,3 @@
 export * from "./util/fs.js";
 export * from "./util/paths.js";
+export * from "./util/alias.js";
